Start of the mig-strip-brutal branch, initially a copy of mig-no280
The idea of this branch is to remove all optimisations to allow a simple
exploration of radically new ideas. Tips #280 and #348 are the first things
to go: they need to know about compiler/executor internals, and hence stand
in the way of big changes to those. The idea is that they will be
reinstated later on.
The branch is based off trunk, but will probably cherrypick from novem*

FossilOrigin-Name: 62c193ec9812887f5a6c22609cce206513577253