All use of the evalFlag value TCL_EVAL_CTX is unused by the code and unreachable by extensions.  This checkin removes all the code supporting that flag value.
The consequence is that all the calls to TclNREvalObjEx() and its callers
that are currently choosing not to pass the TCL_EVAL_DIRECT flag in when
they pass in a non-NULL invoker will no longer be free to change their
mind.  That might be reason not to adopt this change.

FossilOrigin-Name: dfb5bac00f3e1c193121d2c034aab7ceaae3c005