*** NON WORKING BUILD ***
This file integrates miniz in the manner in which the developers of miniz 
intended. However, there are key data structures an capabilities the library
is missing. (Specificially transforming a channel.)

This branch is for posterity. Though if anyone can salvage the effort, have at it.

FossilOrigin-Name: 4f336a73228ccfd5d3eaa645c7ba91c7138778f6