Candidate fix for [e87bcf819f]. Makes sure the unset of a scalar variable linked to an array element terminates searches on that array. In order to accomplish this, some deep revisions in Var handling were done and new routine "Var * TclArrayContaining(Var *varPtr)" was created to return the (Var *) of the array that contains the (Var *) varPtr argument when it is an array element. Callers must verify the argument.
The only code that was unhappy with these deep revisions was the
varname method of TclOO, but revising it to the new situation was not too
difficult. Arguably the code is better now, if a bit uglier and fragile to
further deep Var tinkering.

Second party review would be a good idea, as well as opinions on whether
a backport to 8.6.* is warranted to fix corresponding bugs in the array
enumeration suite of commands.

FossilOrigin-Name: 701d26870f3360783ea3b2b19ce2bf2297b2f32d386456b7bb8a97765ab04b3d